pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sap-llm"
version = "1.0.0"
description = "100% Autonomous Document Processing System for SAP - Zero 3rd Party LLM Dependencies"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "QorSync AI Team", email = "ai@qorsync.com"}
]
keywords = [
    "llm",
    "document-processing",
    "sap",
    "ocr",
    "autonomous-ai",
    "process-automation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML frameworks
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "datasets>=2.15.0",
    "evaluate>=0.4.1",

    # Deep learning optimization
    "deepspeed>=0.12.0",
    "bitsandbytes>=0.41.0",
    "peft>=0.7.0",
    "optimum>=1.15.0",

    # Document processing
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.1",
    "pillow>=10.1.0",
    "layoutparser>=0.3.4",

    # OCR engines
    "easyocr>=1.7.0",

    # NLP utilities
    "sentence-transformers>=2.2.2",
    "spacy>=3.7.0",

    # Graph databases
    "gremlinpython>=3.7.0",
    "neo4j>=5.14.0",
    "networkx>=3.2.0",

    # Databases
    "redis>=5.0.1",
    "pymongo>=4.6.0",
    "motor>=3.3.2",

    # Azure SDKs
    "azure-cosmos>=4.5.1",
    "azure-storage-blob>=12.19.0",
    "azure-servicebus>=7.11.4",
    "azure-identity>=1.15.0",

    # API frameworks
    "fastapi>=0.105.0",
    "uvicorn[standard]>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Async utilities
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    "aiofiles>=23.2.1",

    # Orchestration
    "dapr>=1.12.0",
    "kafka-python>=2.0.2",

    # Cryptography
    "cryptography>=41.0.7",

    # Data validation
    "jsonschema>=4.20.0",
    "pyyaml>=6.0.1",

    # Monitoring & observability
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "tqdm>=4.66.1",
    "numpy>=1.24.0",
    "pandas>=2.1.4",
    "scikit-learn>=1.3.2",

    # Clustering
    "hdbscan>=0.8.33",
    "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "ipython>=8.18.1",
    "jupyter>=1.0.0",
]

gpu = [
    "faiss-gpu>=1.7.4",
    "onnxruntime-gpu>=1.16.3",
]

all = ["sap-llm[dev,gpu]"]

[project.scripts]
sap-llm = "sap_llm.cli:main"
sap-llm-train = "sap_llm.training.train:main"
sap-llm-serve = "sap_llm.api.server:main"

[project.urls]
Homepage = "https://github.com/qorsync/sap-llm"
Documentation = "https://docs.qorsync.com/sap-llm"
Repository = "https://github.com/qorsync/sap-llm"

[tool.setuptools]
packages = ["sap_llm"]

[tool.setuptools.package-data]
sap_llm = ["**/*.json", "**/*.yaml", "**/*.yml"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=sap_llm --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["sap_llm"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
